Fix int32 overflow in presence detection np.ptp() call
np.ptp() computes max - min on the raw int32 signal array. When the
signal spans more than 2^31 (~2.1 billion ADC units), the subtraction
overflows int32 and returns a large negative number. This causes
signal_range > 200_000 to evaluate False, incorrectly marking the user
as not present for that second.

Empirical data from a Pod 5 session shows this affects ~5% of in-bed
records. Fix: cast to int64 before computing ptp so the subtraction
cannot overflow.

+++ biometrics/stream/biometric_processor.py
@@ -169,7 +169,7 @@ class BiometricProcessor:
                 time.sleep(retry_delay)
 
     def detect_presence(self, signal: np.ndarray):
-        signal_range = np.ptp(signal)
+        signal_range = np.ptp(signal.astype(np.int64))
         if signal_range > 200_000:
             self.not_present_for = 0
             self.present_for = self.present_for + 1
